App version 0.3.1
Not ready for release / use on stream! Fixed a few bugs: – Changed lingering "10" in multiplier slide in config form to the correct "100" – The application used Random incorrectly, so notably, playlists were not truly random – When Replay Avoidance had to be disabled because of too few BRBs, the add reasons would be incorrect – A few Hob emotes were coded into the app with incorrect filename extensions – Navigating the forms with [Tab] did not make any sense – Some minor typos.