App version 1.1beta
Not quite ready for release / use on stream. Streamlined the Initial Setup process. In particular, this does not import old BRB playback data anymore. This would not make sense anyway since BRBs needed to be reworked completely. Also, improved handling of the video scrubber and added a button to the main form to switch playback screens.